Add support for Cypher notifications
- Some refactoring done.
