+++ lib/result.js
@@ -18,7 +18,7 @@
  */
  
 import 'babel/polyfill';
-import neo4j from './neo4j';
+import {ResultSummary} from './result-summary';
 
 /**
   * A Result instance is used for retrieving request response.
@@ -110,208 +110,6 @@ class Result {
   }
 }
 
-/**
-  * A ResultSummary instance contains structured metadata for a {Result}.
-  * @access public
-  */
-class ResultSummary {
-  /**
-   * @constructor
-   * @param {string} statement - The statement this summary is for
-   * @param {Object} parameters - Parameters for the statement
-   * @param {Object} metadata - Statement metadata
-   */
-  constructor(statement, parameters, metadata) {
-    let {type: statementType, stats: statistics, plan, profile} = metadata;
-    this.statement = {text: statement, parameters};
-    this.statementType = statementType;
-    this.updateStatistics = new StatementStatistics(statistics || {});
-    this.plan = plan || profile ? new Plan(plan || profile) : false;
-    this.profile = profile ? new ProfiledPlan(profile) : false;
-  }
-
-  /**
-   * Check if the result summary has a plan
-   * @return {boolean}
-   */
-  hasPlan() {
-    return this.plan instanceof Plan
-  }
-
-  /**
-   * Check if the result summary has a profile
-   * @return {boolean}
-   */
-  hasProfile() {
-    return this.profile instanceof ProfiledPlan
-  }
-}
-
-/**
-  * Class for execution plan received by prepending Cypher with EXPLAIN.
-  * @access public
-  */
-class Plan {
-  /**
-   * Create a Plan instance
-   * @constructor
-   * @param {Object} plan - Object with plan data
-   */
-  constructor(plan) {
-    this.operatorType = plan.operatorType;
-    this.identifiers = plan.identifiers;
-    this.arguments = plan.args;
-    this.children = plan.children ? plan.children.map((child) => new Plan(child)) : [];
-  }
-}
-
-/**
-  * Class for execution plan received by prepending Cypher with PROFILE.
-  * @access public
-  */
-class ProfiledPlan {
-  /**
-   * Create a ProfiledPlan instance
-   * @constructor
-   * @param {Object} plan - Object with plan data
-   */
-  constructor(profile) {
-    this.operatorType = profile.operatorType;
-    this.identifiers = profile.identifiers;
-    this.arguments = profile.args;
-    this.dbHits = profile.args.DbHits.toInt();
-    this.rows = profile.args.Rows.toInt();
-    this.children = profile.children ? profile.children.map((child) => new ProfiledPlan(child)) : [];
-  }
-}
-
-/**
-  * Get statistical information for a {Result}.
-  * @access public
-  */
-class StatementStatistics {
-  /**
-   * Structurize the statistics
-   * @constructor
-   * @param {Object} statistics - Result statistics
-   */
-  constructor(statistics) {
-    this._stats = {
-      nodesCreated: 0,
-      nodesDelete: 0,
-      relationshipsCreated: 0,
-      relationshipsDeleted: 0,
-      propertiesSet: 0,
-      labelsAdded: 0,
-      labelsRemoved: 0,
-      indexesAdded: 0,
-      indexesRemoved: 0,
-      constraintsAdded: 0,
-      constraintsRemoved: 0
-    }
-    Object.keys(statistics).forEach((index) => {
-      let val = neo4j.isInt(statistics[index]) ? statistics[index].toInt() : statistics[index];
-      //To camelCase
-      this._stats[index.replace(/(\-\w)/g, (m) => m[1].toUpperCase())] = val;
-    });
-  }
-
-  /**
-   * Did the database get updated?
-   * @return {boolean}
-   */
-  containsUpdates() {
-    return Object.keys(this._stats).reduce((last, current) => {
-      return last + this._stats[current];
-    }, 0) > 0;
-  }
-
-  /**
-   * @return {Number} - Number of nodes created.
-   */
-  nodesCreated() {
-    return this._stats.nodesCreated;
-  }
-
-  /**
-   * @return {Number} - Number of nodes deleted.
-   */
-  nodesDeleted() {
-    return this._stats.nodesDeleted;
-  }
-
-  /**
-   * @return {Number} - Number of relationships created.
-   */
-  relationshipsCreated() {
-    return this._stats.relationshipsCreated;
-  }
-
-  /**
-   * @return {Number} - Number of nodes deleted.
-   */
-  relationshipsDeleted() {
-    return this._stats.relationshipsDeleted;
-  }
-
-  /**
-   * @return {Number} - Number of properties set.
-   */
-  propertiesSet() {
-    return this._stats.propertiesSet;
-  }
-
-  /**
-   * @return {Number} - Number of labels added.
-   */
-  labelsAdded() {
-    return this._stats.labelsAdded;
-  }
-
-  /**
-   * @return {Number} - Number of labels removed.
-   */
-  labelsRemoved() {
-    return this._stats.labelsRemoved;
-  }
-
-  /**
-   * @return {Number} - Number of indexes added.
-   */
-  indexesAdded() {
-    return this._stats.indexesAdded;
-  }
-
-  /**
-   * @return {Number} - Number of indexes removed.
-   */
-  indexesRemoved() {
-    return this._stats.indexesRemoved;
-  }
-
-  /**
-   * @return {Number} - Number of contraints added.
-   */
-  constraintsAdded() {
-    return this._stats.constraintsAdded;
-  }
-
-  /**
-   * @return {Number} - Number of contraints removed.
-   */
-  constraintsRemoved() {
-    return this._stats.constraintsRemoved;
-  }
-}
-
-const statementType = {
-  READ_ONLY: 'r',
-  READ_WRITE: 'rw',
-  WRITE_ONLY: 'w',
-  SCHEMA_WRITE: 's'
-}
-
 export default {
-  Result,
-  statementType
+  Result
 }
